tetragon/windows: Cimpilation only change to compile cgroups package
Add windows specific stub function definitions to compile pkg/cgroups package on Windows.
Common types and functions added in cgroups.go
